many changes

+++ src/main/groovy/com/fmillone/fci/fundStatus/TrustStatus.groovy
@@ -20,7 +20,7 @@ class TrustStatus {
     String Horiz
     LocalDate date
     Long amountOfPieces
-    Long totalValue
+    Double totalValue
     Double valuesPerUnity
